requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Sock==0.7.0
Werkzeug==2.3.7
mutagen==1.47.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
mypy==1.5.1
python-dateutil==2.8.2
pydantic==2.3.0
ffmpeg-python==0.2.0
redis==5.0.0
celery==5.3.4
APScheduler==3.10.4
psutil==5.9.5
typing-extensions==4.7.1
aiohttp==3.8.5
asyncio==3.4.3
colorama==0.4.6
tqdm==4.66.1
PyYAML==6.0.1
cryptography==41.0.3
bcrypt==4.0.1
PyJWT==2.8.0
Pillow==10.0.0
numpy==1.25.2
pandas==2.1.0
matplotlib==3.7.2
seaborn==0.12.2
